sempre usar o inteiro(int) padrão
De uma vez por todas se livrando deste erro aqui. Incrível como sempre volta.

+++ include/teste.h
@@ -5,22 +5,22 @@
 #include <stdbool.h>
 #include <stdint.h>
 #include "terminal.h"
-   
-// nome mais curto, estético e significativo para um ponteiro de 
+
+// nome mais curto, estético e significativo para um ponteiro de
 // de função.
  typedef void(*Fn)(void);
  // Tipo de função especial para testes-unitários.
  typedef int (*FnTst) (void);
 
-/* string vázia dizendo que não quer que o teste tem um nome própria, ou 
+/* string vázia dizendo que não quer que o teste tem um nome própria, ou
  * uma mensagem explicando-o. */
  #define SEM_MENSAGEM ""
  #define SEM_NOME SEM_MENSAGEM
 
  extern void debug_aqui(void);
 /* Executa variados testes passados. Os parâmetros seguem a seguinte
- * lógica: o total de argumentos totais, aí começa a alternância 
- * entre o endereço da funçõa a ser executada, a confirmação se você 
+ * lógica: o total de argumentos totais, aí começa a alternância
+ * entre o endereço da funçõa a ser executada, a confirmação se você
  * deseja realmente executa-lô; um bocado de função seguida do valor
  * verdade da sua execução, descotando o primeiro, que é o total de
  * funções e valores-verdades. Exemplo:
@@ -29,19 +29,19 @@
  *
  * Lembre-se que todos pares tem que bater com a seguinte função e
  * valor verdade, além de bate com o total de argumentos declarados
- * inicialmente. 
+ * inicialmente.
  *
  *   Cada nova geração ganha uma letra mais avançada no alfabeto, isso até
  * este tipo de teste for realmente substituído por algo mais avançado. */
-void executa_tst(char*, Fn, bool); 
-void executa_testes(uint8_t total, ...);
-void executa_testes_a(bool ativado, int total, ...);
+void executa_tst       (char*, Fn, bool);
+void executa_testes    (int total, ...);
+void executa_testes_a  (bool ativado, int total, ...);
 
-/* Converte um valor lógico para uma string, em português, dizendo o que 
- * ele representa. Nome similires, mas fazem a mesma coisas também estão 
- * dados abaixo, por motivos de compatibilidade. A outra função faz 
- * exatamente o reverso, pega uma string que representa um valor verdade, 
- * então retorna o valor-verdade em sí. 
+/* Converte um valor lógico para uma string, em português, dizendo o que
+ * ele representa. Nome similires, mas fazem a mesma coisas também estão
+ * dados abaixo, por motivos de compatibilidade. A outra função faz
+ * exatamente o reverso, pega uma string que representa um valor verdade,
+ * então retorna o valor-verdade em sí.
  */
  extern bool str_to_bool(char* s);
